Move hooks into different file to get rid of HMR warnings
