+++ src/packages/scale-viewer/ScaleViewer.test.tsx
@@ -3,7 +3,7 @@
 /**
  * @vitest-environment jsdom
  */
-import { MapContainer, OlMapConfigurationProvider } from "@open-pioneer/experimental-ol-map";
+import { OlMapConfigurationProvider } from "@open-pioneer/experimental-ol-map";
 import { OlMapRegistry } from "@open-pioneer/experimental-ol-map/services";
 import { Service, ServiceOptions } from "@open-pioneer/runtime";
 import {
@@ -11,13 +11,10 @@ import {
     PackageContextProviderProps
 } from "@open-pioneer/test-utils/react";
 import { createService } from "@open-pioneer/test-utils/services";
-import { act, render, renderHook, screen, waitFor } from "@testing-library/react";
+import { render, screen, waitFor } from "@testing-library/react";
 import { MapOptions } from "ol/Map";
 import { expect, it } from "vitest";
-import { ScaleViewer, useCenter, useResolution, useScale } from "./ScaleViewer";
-import View from "ol/View";
-import OSM from "ol/source/OSM";
-import TileLayer from "ol/layer/Tile";
+import { ScaleViewer } from "./ScaleViewer";
 
 // used to avoid a "ResizeObserver is not defined" error
 global.ResizeObserver = require("resize-observer-polyfill");
@@ -132,209 +129,4 @@ it("should successfully create a scale viewer component with additional css clas
         const styles = window.getComputedStyle(box);
         expect(styles.paddingLeft).toBe("1px");
     }
-});
-
-it("should successfully create a map resolution", async () => {
-    const mapId = "test";
-    const zoom = 10;
-    const mapOptions = {
-        view: new View({
-            projection: "EPSG:3857",
-            center: [847541, 6793584],
-            zoom
-        }),
-        layers: [
-            new TileLayer({
-                source: new OSM(),
-                properties: { title: "OSM" }
-            })
-        ]
-    } as MapOptions;
-    const service = await createOlMapRegistry(mapId, mapOptions);
-
-    render(
-        <PackageContextProvider {...createPackageContextProviderProps(service)}>
-            <div data-testid="base">
-                <MapContainer mapId={mapId} />
-                <ScaleViewer mapId={mapId}></ScaleViewer>
-            </div>
-        </PackageContextProvider>
-    );
-
-    // assert map and scale viewer is mounted
-    await waitFor(async () => {
-        const domElement = await screen.findByTestId("base");
-        const container = domElement.querySelector(".ol-viewport");
-        if (!container) {
-            throw new Error("map not mounted");
-        }
-        return domElement;
-    });
-
-    const map = await service.getMap(mapId);
-    if (!map) {
-        throw new Error("map not defined");
-    }
-
-    const view = map.getView();
-    if (!view) {
-        throw new Error("view not defined");
-    }
-
-    const mapResolution = view.getResolution();
-    if (!mapResolution) {
-        throw new Error("resolution not defined");
-    }
-
-    let mapZoom = view.getZoom();
-
-    // change zoom level and detect resolution change
-    const hook = renderHook(() => useResolution(map));
-    const result = hook.result;
-
-    const firstResolution = result.current.resolution;
-    expect(firstResolution).not.toBe(undefined);
-
-    await act(async () => {
-        if (!mapZoom) {
-            throw new Error("zoom not defined");
-        }
-
-        view.setZoom(++mapZoom);
-        view.dispatchEvent("change:resolution");
-    });
-    hook.rerender();
-
-    const nextResolution = hook.result.current.resolution;
-    expect(firstResolution).not.toBe(nextResolution);
-});
-
-it("should successfully create a map center", async () => {
-    const mapId = "test";
-    const zoom = 10;
-    const center = [847541, 6793584];
-    const mapOptions = {
-        view: new View({
-            projection: "EPSG:3857",
-            center,
-            zoom
-        }),
-        layers: [
-            new TileLayer({
-                source: new OSM(),
-                properties: { title: "OSM" }
-            })
-        ]
-    } as MapOptions;
-    const service = await createOlMapRegistry(mapId, mapOptions);
-
-    render(
-        <PackageContextProvider {...createPackageContextProviderProps(service)}>
-            <div data-testid="base">
-                <MapContainer mapId={mapId} />
-                <ScaleViewer mapId={mapId}></ScaleViewer>
-            </div>
-        </PackageContextProvider>
-    );
-
-    // assert map and scale viewer is mounted
-    await waitFor(async () => {
-        const domElement = await screen.findByTestId("base");
-        const container = domElement.querySelector(".ol-viewport");
-        if (!container) {
-            throw new Error("map not mounted");
-        }
-        return domElement;
-    });
-
-    const map = await service.getMap(mapId);
-    if (!map) {
-        throw new Error("map not defined");
-    }
-
-    const view = map.getView();
-    if (!view) {
-        throw new Error("view not defined");
-    }
-
-    // change center and detect center change
-    const hook = renderHook(() => useCenter(map));
-    const result = hook.result;
-
-    const firstCenter = result.current.center;
-    expect(firstCenter).not.toBe(undefined);
-
-    await act(async () => {
-        view.setCenter([1489200, 6894026]);
-        view.dispatchEvent("change:center");
-    });
-    hook.rerender();
-
-    const nextCenter = hook.result.current.center;
-    expect(firstCenter).not.toBe(nextCenter);
-});
-
-it("should successfully create a map scale", async () => {
-    const mapId = "test";
-    const scale = 336409;
-    const zoom = 10;
-    const center = [847541, 6793584];
-    const mapOptions = {
-        view: new View({
-            projection: "EPSG:3857",
-            center,
-            zoom
-        }),
-        layers: [
-            new TileLayer({
-                source: new OSM(),
-                properties: { title: "OSM" }
-            })
-        ]
-    } as MapOptions;
-    const service = await createOlMapRegistry(mapId, mapOptions);
-
-    render(
-        <PackageContextProvider {...createPackageContextProviderProps(service)}>
-            <div data-testid="base">
-                <MapContainer mapId={mapId} />
-                <ScaleViewer mapId={mapId}></ScaleViewer>
-            </div>
-        </PackageContextProvider>
-    );
-
-    // assert map and scale viewer is mounted
-    await waitFor(async () => {
-        const domElement = await screen.findByTestId("base");
-        const container = domElement.querySelector(".ol-viewport");
-        if (!container) {
-            throw new Error("map not mounted");
-        }
-        return domElement;
-    });
-
-    const map = await service.getMap(mapId);
-    if (!map) {
-        throw new Error("map not defined");
-    }
-
-    const view = map.getView();
-    if (!view) {
-        throw new Error("view not defined");
-    }
-
-    const mapResolution = view.getResolution();
-    if (!mapResolution) {
-        throw new Error("resolution not defined");
-    }
-
-    const mapCenter = view.getCenter();
-    if (!mapCenter) {
-        throw new Error("center not defined");
-    }
-
-    // get map scale
-    const hook = renderHook(() => useScale(map, mapResolution, mapCenter));
-    const result = hook.result;
-    expect(result.current.scale).toBe(scale);
 });
